using_streamlit/requirements.txt
streamlit 
google-generativeai
langchain-google-genai
langchain
langchain_community
python-dotenv
faiss-cpu
PyPDF2
qdrant-client
python-docx